# docs/openapi.yaml
openapi: 3.1.0
info:
  description: "로그인, 토큰 저장, 401 처리처럼 여러 호출에 걸친 흐름은 tyt-be의 docs/api-guide.md를 본다."
  title: TYT API
  version: v1
servers:
- description: 로컬
  url: http://localhost:8080
security:
- bearerAuth: []
tags:
- description: 카카오 로그인과 토큰 재발급. 토큰 없이 호출한다.
  name: Auth
- description: 사용자
  name: User
paths:
  /api/v1/auth/kakao/tokens:
    post:
      description: |-
        앱이 카카오 SDK 로그인으로 받은 카카오 액세스 토큰으로 로그인한다. 처음 보는 카카오 계정이면 가입까지 한다.
        응답의 accessToken은 이후 요청의 `Authorization: Bearer` 헤더에, refreshToken은 재발급에 쓴다. 두 토큰을 모두 저장한다.
      operationId: loginWithKakao
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KakaoLoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTokenResponse"
          description: '`LOGIN_SUCCESS`'
        "400":
          content:
            application/json:
              example:
                code: INVALID_REQUEST
                message: "accessToken: 공백일 수 없습니다"
                data: null
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: "`INVALID_REQUEST`: accessToken이 비어 있다"
        "401":
          content:
            application/json:
              example:
                code: INVALID_SOCIAL_TOKEN
                message: 소셜 로그인 토큰이 유효하지 않습니다.
                data: null
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: "`INVALID_SOCIAL_TOKEN`: 카카오 토큰이 만료됐거나 유효하지 않거나 다른 앱에서 발급됐다\
            . 카카오 로그인부터 다시 한다."
      security: []
      summary: 카카오 로그인
      tags:
      - Auth
  /api/v1/auth/tokens:
    post:
      description: |-
        refresh 토큰으로 access와 refresh를 모두 새로 발급한다. 쓴 refresh 토큰은 바로 무효가 되므로 응답의 두 토큰으로 교체해 저장한다.
        같은 refresh 토큰으로 동시에 두 번 호출하면 하나만 성공하고 나머지는 401이다. 재발급 요청은 한 곳에서 한 번에 하나만 보낸다.
      operationId: reissue
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReissueTokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTokenResponse"
          description: '`TOKEN_REISSUED`'
        "400":
          content:
            application/json:
              example:
                code: INVALID_REQUEST
                message: "refreshToken: 공백일 수 없습니다"
                data: null
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: "`INVALID_REQUEST`: refreshToken이 비어 있다"
        "401":
          content:
            application/json:
              example:
                code: INVALID_REFRESH_TOKEN
                message: refresh 토큰이 유효하지 않습니다. 다시 로그인해야 합니다.
                data: null
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: "`INVALID_REFRESH_TOKEN`: refresh 토큰이 만료·무효이거나, 이미 교체됐거나, 다\
            른 기기에서 로그인해 무효가 됐다. 저장된 토큰을 지우고 로그인 화면으로 보낸다."
      security: []
      summary: 토큰 재발급
      tags:
      - Auth
  /api/v1/users/me:
    get:
      description: access 토큰의 사용자를 조회한다. 앱 시작 시 저장된 토큰이 아직 쓸 수 있는지 확인하는 용도로도 쓴다.
      operationId: getMe
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"
          description: '`USER_FOUND`'
        "401":
          content:
            application/json:
              example:
                code: UNAUTHENTICATED
                message: 인증이 필요합니다.
                data: null
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: "`UNAUTHENTICATED`: 토큰이 없거나 만료·무효다. 재발급 후 다시 호출한다."
        "404":
          content:
            application/json:
              example:
                code: USER_NOT_FOUND
                message: 사용자를 찾을 수 없습니다.
                data: null
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: "`USER_NOT_FOUND`: 토큰은 유효하지만 사용자가 없다. 저장된 토큰을 지우고 로그인 화면으로\
            \ 보낸다."
      summary: 내 정보 조회
      tags:
      - User
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          description: 성공·실패 코드. 화면 분기는 이 값으로 한다
          example: LOGIN_SUCCESS
        data:
          description: 성공 시 응답 본문. 실패 시 null
        message:
          type: string
          description: 사람이 읽는 메시지. 문구가 바뀔 수 있으니 분기에 쓰지 않는다
          example: 로그인했습니다.
    ApiResponseTokenResponse:
      type: object
      properties:
        code:
          type: string
          description: 성공·실패 코드. 화면 분기는 이 값으로 한다
          example: LOGIN_SUCCESS
        data:
          $ref: "#/components/schemas/TokenResponse"
          description: 성공 시 응답 본문. 실패 시 null
        message:
          type: string
          description: 사람이 읽는 메시지. 문구가 바뀔 수 있으니 분기에 쓰지 않는다
          example: 로그인했습니다.
    ApiResponseUserResponse:
      type: object
      properties:
        code:
          type: string
          description: 성공·실패 코드. 화면 분기는 이 값으로 한다
          example: LOGIN_SUCCESS
        data:
          $ref: "#/components/schemas/UserResponse"
          description: 성공 시 응답 본문. 실패 시 null
        message:
          type: string
          description: 사람이 읽는 메시지. 문구가 바뀔 수 있으니 분기에 쓰지 않는다
          example: 로그인했습니다.
    KakaoLoginRequest:
      type: object
      properties:
        accessToken:
          type: string
          description: 카카오 SDK 로그인으로 받은 카카오 액세스 토큰
          example: kakao-sdk-access-token
          minLength: 1
      required:
      - accessToken
    ReissueTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: 로그인이나 직전 재발급에서 받은 refreshToken
          example: eyJhbGciOiJIUzI1NiJ9.refresh
          minLength: 1
      required:
      - refreshToken
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "API 호출에 쓰는 JWT. `Authorization: Bearer`에 넣는다. 유효기간 30분"
          example: eyJhbGciOiJIUzI1NiJ9.access
        refreshToken:
          type: string
          description: 재발급에만 쓰는 JWT. 유효기간 14일. 재발급하면 새 값으로 바뀌고 이전 값은 무효
          example: eyJhbGciOiJIUzI1NiJ9.refresh
    UserResponse:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
          description: 가입 시각. 한국 시간(Asia/Seoul) 기준이며 오프셋 없는 ISO-8601이다
          example: 2026-09-14T00:00:00
        id:
          type: integer
          format: int64
          description: 사용자 id
          example: 1
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: POST /api/v1/auth/kakao/tokens 또는 POST /api/v1/auth/tokens로 받은
        accessToken
      scheme: bearer
      type: http